discovery/kubernetes: extract node discovery
This change extracts node discovery into its own type.
